Details in the commit message
1. Removed the '\"" characters from MQTT_PORT macro (it shouldn't
have according to the protocol)
2. Fixed a typo in MQTT_PASSWORD (it was MQTT_PSSWORD)
3. Changing the COMM_STATE GET_MAC_ADD to GET_MAC_ADDR
4. Encapsulating the sending message function (comm_send_msg now)
5. Implementing the GET_IP_NUMB and GET_MAC_ADDR in the
comm_fsm_recv function
(Also checking if we already have the IP number and the MAC addres.
Since they won't change, we don't need to get'em again. In other
words, if we already have the IP number and the MAC address, jumps
from CONNECT_WIFI to CONNECT_MQTT.)
6. Implementing the GET_IP_NUMB, GET_MAC_ADDR and CONNECT_MQTT
state in the comm_fsm_send functon.