Properly event-driven architecture with event bufferization.
1. A ring buffer was implemented in order to bufferize the events.
Also a event handler function was implemented (actually it is just
an encapsulation of the handlers that were already implemented
before). The event NEW_TEMPERATURE_MEAS has not been handled yet.

2. The UARTx.[c|h] - generated by PE - had to be changed since I
needed a function to check if a message was completely received
without removing the character from the UARTx RX ring buffers. This
is also the reason why I'm pushing the generated code.